% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggLsys.R
\name{ggLsys}
\alias{ggLsys}
\title{S3 method to draw a 2D L-System with ggplot2}
\usage{
ggLsys(x, ...)
}
\arguments{
\item{x}{A character vector, or dataframe containing all required info to plot Lsys. For
details, see \code{\link{drawLsys}}}
}
\value{
a gg object, by default it would be plotted.
}
\description{
This function takes either returned dataframe from Lsys from \code{\link{recLsys}}
or initial strings to create which, to create plots with ggplot
}
\keyword{plot}
