% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/LindenmayeR-package.R
\docType{package}
\name{LindenmayeR-package}
\alias{LindenmayeR}
\alias{LindenmayeR-package}
\title{Lindenmayer System Functions}
\description{
Functions to Explore L-Systems (Lindenmayer Systems)
}
\details{
Lindenmayer or L-systems are parallel rewriting systems
which can be used to simulate biological forms and certain
kinds of fractals.  Briefly, in an L-system a series of
symbols in a string are replaced iteratively according to
rules to give a more complex string. Eventually, the
symbols are translated into turtle
graphics for plotting.  Wikipedia has a very good
introduction: \url{https://en.wikipedia.org/wiki/L-system}
This package provides basic functions for exploring L-systems.
}
\author{
Bryan A. Hanson
}
\keyword{package}

