% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recLsys.R
\name{recLsys}
\alias{recLsys}
\title{Record path of the virtual pen with turtle graphics from Lsys}
\usage{
recLsys(string, drules = NULL, st = c(5, 50, 0), stepSize = 1,
  ang = 90, which = length(string), shrinkFactor = NULL)
}
\arguments{
\item{string}{A character vector, or dataframe containing all required info to plot Lsys. For
details, see \code{\link{drawLsys}}}
}
\value{
a data frame, containing coordinate data and segment types
  s for start of a segment d for segment endpoints to be drawn, r for reverse points#'
}
\description{
This function takes input strings, previously created with \code{\link{Lsys}},
translates them to coordinates and types of line segments stored in a data frame.
}
\keyword{coordinate}
